requirements.txt
django == 3.2.6
djangorestframework == 3.12.4
django-filter == 2.4.0
markdown == 3.3.4
confuse == 1.5.0
django-guardian == 2.4.0